% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{geogit_location}
\alias{geogit_location}
\title{Import Github users from a specific location}
\usage{
geogit_location(location, token, page = 1, sort = "followers")
}
\arguments{
\item{location}{Desired location to explore e.g. Morocco}

\item{token}{The geogit_token that contains access token information.
If not provided, the anonymous user token is used.}

\item{page}{Page number of search results. Maximum value is 10.}

\item{sort}{Value to sort followers by. You can sort users by
\code{followers}, \code{repositories}, or \code{joined} (account creation timestamp).}
}
\value{
A data frame that contains the users that match the given location.
}
\description{
This function retrieves Github users from a specific location
}
\details{
The first 100 users who match the location are returned.
By default, the first page of the search is returned and the users are sorted by number of
followers. Each search page has a maximum of 100 search results, and
you can access the other search pages by changing the page parameter.
Github API limits the number of searches that can be retrieved to 1000 (10 pages).
Only users with at least 1 repository on Github are retrieved using this function.
This function is particulatly useful for retrieving the usernames that match a location.
To retrieve more information about a user, you can use \code{geogit_user}.
}
\examples{
geogit_location("Morocco")
geogit_location("Morocco", page = 2) #to retrieve the next 100 search results
geogit_location("Morocco", sort = "repositories") #to sort by repositories instead of followers

}
