% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remaining_requests.R
\name{geogit_remaining_requests}
\alias{geogit_remaining_requests}
\title{Retrieve the number of remaining Github API requests available}
\usage{
geogit_remaining_requests(token)
}
\arguments{
\item{geogit_token}{The geogit_token used to retrieve data from Github}
}
\value{
A dataframe with information about the number of requests available
for each component of the Github API.
}
\description{
Retrieve the number of remaining Github API requests available
}
\examples{
token <- geogit_token()
geogit_remaining_requests(token)

}
