% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{geogit_repos}
\alias{geogit_repos}
\title{Retrieve information about a Github user's repositories.}
\usage{
geogit_repos(github_username, token, page = 1, sort = "created")
}
\arguments{
\item{github_username}{The Github user's username on the platform.}

\item{token}{The geogit_token that contains access token information.}

\item{page}{The page number of the repository results returned by the API}

\item{sort}{The parameter to sort the repositories by. You can sort repositories
by date created \code{created}, date updated \code{updated}, or date pushed \code{pushed}.}
}
\value{
A tibble that contains the given user's repository information.
}
\description{
This function retrieves information about a user's public repositories on GitHub.
}
\details{
All pieces of information provided by the Github API about the repositories are returned.
Only the first 100 repositories from the API response will be returned. Use the page parameter
to access different pages of the repositories API response. Each page can have a maximum
of 100 results. Iterate through multiple pages to retrieve all repositories.
}
\examples{
token <- geogit_token()
geogit_repos("abelemlih", token)

}
