% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{geogit_token}
\alias{geogit_token}
\title{Create a token to access GitHub API}
\usage{
geogit_token(github_access_token)
}
\arguments{
\item{github_access_token}{The string from your github settings ... blah blah
If not provided, an anonymous token is generated.}
}
\value{
A token object for use in other geogit_ functions.
}
\description{
You need a token because ...
This can be as long as needed.
}
\details{
With an anonymous token, your access is limited to ...
To get a token, you need a github account: See vignette.
}
\examples{
geogit_token("Something that looks right.")

}
