% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{geogit_user}
\alias{geogit_user}
\title{Import user information from Github}
\usage{
geogit_user(github_user, token = "")
}
\arguments{
\item{token}{geogit_token that contains access token information.
If not provided, the anonymous user token is used.}

\item{github_username}{Github user's username on the platform.}
}
\value{
A data frame that contains the given user's information.
}
\description{
This function retrieves a Github user's public information
}
\details{
#' All pieces of information provided by the Github API about the user are returned.
}
\examples{
geogit_user("abelemlih")

}
