% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{geogit_token}
\alias{geogit_token}
\title{Create a token to access GitHub API}
\usage{
geogit_token(github_access_token = "")
}
\arguments{
\item{github_access_token}{The personal access token from your Github
account's developer settings. If not provided, an unauthenticated token is generated.}
}
\value{
A token \code{list} for use in other geogit_ functions.
}
\description{
You need a token to authenticate and access the Github API
through your Github account.
}
\details{
Authenticated account can process up to 5000 requests per hour
and 30 search requests per minute, instead of 60 requests per hour and 10 search
request per minute without authentication. For more information,
see https://developer.github.com/v3/#rate-limiting. To get a token,
you need a github account (See vignette).
}
\examples{
geogit_token("personal_acess_token")

}
